backend/requirements.txt
fastapi>=0.100.0,<0.110.0
uvicorn>=0.24.0,<0.25.0
sqlalchemy>=2.0.0,<3.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.7,<0.1.0
langchain>=0.1.0,<0.2.0
pytest>=8.0.0,<9.0.0
httpx>=0.26.0,<0.27.0
openai>=1.0.0,<2.0.0
alembic>=1.13.0,<2.0.0